Refactor out Events into its own utility class
